requirements.txt
Pillow==8.3.2
qrcode[pil]==7.3
python-docx==0.8.11